Update README.md
The 2005 movie ‘Coach Carter’ is a biographical teen sports drama about the incredible basketball coach Ken Carter.

Coach Carter is hired to be the new basketball coach for the Richmond Oilers. Coach Ken Carter after becoming the new coach he requires the team members if they want to play in his team, they have to sign a contract the contract that state that they have attended all their classes and sit in the front row of all their classes and maintain a C+. Coach Carter is a very disciplined type of basketball coach. He makes his players, if they are late for training, run, and if they are disrespectful, they will do push-ups. During practice he teaches the team the importance of respect and discipline throughout practices, he uses names of women in his life as playing tactics, and his team become undefeated through hard work.

Most of it is set in a part of town where there’s high crime rates. There are scenes set around the basketball court and school. Some of their houses are included in the inner-city ghetto setting, suggesting that the characters are not very rich. One of the characters was giving money to his cousin who is in a gang in the ghetto part of Richmond.

After getting progress reports from his players and finding out some of them have broken the terms of the contract, Coach Carter benches his undefeated team and locks the basketball court and cancels their games for poor grades in school. He sends all of them to the library to fix their grades with the help of three teachers who volunteered their time. The patents of the players get tired of the lockout and bring the matter to the school board and the outcome of that was voted to end the lockout. Coach Carter quits his job because he can’t support a school that cares more about basketball over good grades. He goes to the basketball court to get his things, but finds his team, studying instead of training. His team reaches their goal and can play and train again. The team gets invited to the state high school playoffs, and ends up losing to their rival St. Francis by just 2 points, but Coach Carter was proud of them for their effort as a team. Six players were offered college and graduate scholarships.

Overall, the movie ‘Coach Carter’ underlined the power of teamwork. It shows how a group of young men went from a team with low team morale to a team with high morale, and together, under the unconventional methods of their coach, they beat the odds. Coach Carter motivated them to become winners. The movie is filled with inspiring and upfront messages about overcoming adversity.